+++ ElvUI_LuckyoneUI/Core/Update.lua
@@ -3,13 +3,25 @@ local L1UI, E, L, V, P, G = unpack(select(2, ...))
 -- Ingame config buttons 'Update Main' & 'Update Healing'
 function L1UI:UpdateLayout(layout)
 
+	-- ActionBar db updates
+	if E.Retail then
+		E.db.actionbar.bar1.professionQuality.enable = false
+		E.db.actionbar.bar2.professionQuality.enable = false
+		E.db.actionbar.bar3.professionQuality.enable = false
+		E.db.actionbar.bar4.professionQuality.enable = false
+		E.db.actionbar.bar5.professionQuality.enable = false
+		E.db.actionbar.bar6.professionQuality.enable = false
+		E.db.actionbar.bar7.professionQuality.enable = false
+		E.db.actionbar.bar8.professionQuality.enable = false
+		E.db.actionbar.bar9.professionQuality.enable = false
+		E.db.actionbar.bar10.professionQuality.enable = false
+		E.db.actionbar.bar13.professionQuality.enable = false
+		E.db.actionbar.bar14.professionQuality.enable = false
+		E.db.actionbar.bar15.professionQuality.enable = false
+	end
+
 	-- Bump install version
 	E.db.L1UI.install_version = L1UI.Version
 
 	L1UI:Print(L["Successfully updated your current layout to LuckyoneUI version "]..L1UI.Version)
-
-	-- ToDo: Remove me before 2.19
-	L1UI:Print('If you are looking for the new NamePlates style please re-apply the profile:')
-	L1UI:Print('For ElvUI NamePlates: in /luckyoneui config -> Profiles tab')
-	L1UI:Print('For Plater: Re-import from https://wago.io/LuckyoneUI-Plater')
 end